+++ packages/desktop/app/components/Sidebar/Sidebar.js
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 import Divider from '@material-ui/core/Divider';
 import MuiList from '@material-ui/core/List';
 
+import { TamanuLogoWhite } from '../TamanuLogo';
 import { version } from '../../../package.json';
 
 import { LogoutItem } from './LogoutItem';
@@ -12,31 +13,16 @@ import { FacilityNameDisplay } from '../FacilityNameDisplay';
 
 import { Colors } from '../../constants';
 
-const SidebarContainer = styled.div`
-  grid-row: 2 / -1;
-  grid-column: 1 / 2;
-
-  min-width: 275px;
-  position: relative;
+const Container = styled.div`
   background: ${Colors.primaryDark};
-  box-shadow: 1px 0px 4px rgba(0, 0, 0, 0.15);
+  min-width: 260px;
+  padding-left: 8px;
+  box-shadow: 1px 0 4px rgba(0, 0, 0, 0.15);
   color: ${Colors.white};
-  flex-grow: 0;
-  flex-shrink: 0;
-
-  display: flex;
-  flex-direction: column;
 
   i {
     color: ${Colors.white};
   }
-`;
-
-const SidebarMenuContainer = styled.div`
-  flex-grow: 1;
-  overflow: auto;
-  display: grid;
-  grid-template-rows: auto 1fr;
 `;
 
 const AdditionalInfo = styled.div`
@@ -48,8 +34,8 @@ const AdditionalInfo = styled.div`
   font-size: 0.8rem;
 `;
 
-const List = styled(MuiList)`
-  padding-top: 0;
+const Logo = styled(TamanuLogoWhite)`
+  margin: 24px 0 10px 20px;
 `;
 
 export class Sidebar extends Component {
@@ -80,41 +66,40 @@ export class Sidebar extends Component {
     const { selectedParentItem } = this.state;
     const { currentPath, items, onPathChanged, permissionCheck = () => true } = this.props;
     return (
-      <SidebarContainer>
-        <SidebarMenuContainer>
-          <List component="nav">
-            {items.map(item => (
-              <PrimarySidebarItem
-                icon={item.icon}
-                label={item.label}
-                key={item.key}
-                selected={selectedParentItem === item.key}
-                onClick={() => this.clickedParentItem(item)}
-              >
-                {item.children.map(child => (
-                  <SecondarySidebarItem
-                    key={child.path}
-                    path={child.path}
-                    isCurrent={currentPath === child.path}
-                    icon={child.icon}
-                    label={child.label}
-                    disabled={!permissionCheck(child, item)}
-                    onClick={() => onPathChanged(child.path)}
-                  />
-                ))}
-              </PrimarySidebarItem>
-            ))}
-          </List>
-          <div>
-            <Divider />
-            <LogoutItem onClick={this.onLogout} />
-          </div>
-        </SidebarMenuContainer>
+      <Container>
+        <Logo size="140px" />
+        <MuiList component="nav">
+          {items.map(item => (
+            <PrimarySidebarItem
+              icon={item.icon}
+              label={item.label}
+              key={item.key}
+              selected={selectedParentItem === item.key}
+              onClick={() => this.clickedParentItem(item)}
+            >
+              {item.children.map(child => (
+                <SecondarySidebarItem
+                  key={child.path}
+                  path={child.path}
+                  isCurrent={currentPath === child.path}
+                  icon={child.icon}
+                  label={child.label}
+                  disabled={!permissionCheck(child, item)}
+                  onClick={() => onPathChanged(child.path)}
+                />
+              ))}
+            </PrimarySidebarItem>
+          ))}
+        </MuiList>
+        <div>
+          <Divider />
+          <LogoutItem onClick={this.onLogout} />
+        </div>
         <AdditionalInfo>
           <div>Version {version}</div>
           <FacilityNameDisplay />
         </AdditionalInfo>
-      </SidebarContainer>
+      </Container>
     );
   }
 }