+++ packages/lan/app/routes/apiv1/patient.js
@@ -1,4 +1,6 @@
 import express from 'express';
+import config from 'config';
+import moment from 'moment';
 import asyncHandler from 'express-async-handler';
 import { QueryTypes, Op } from 'sequelize';
 import { isEqual } from 'lodash';
@@ -11,6 +13,7 @@ import { createPatientFilters } from '../../utils/patientFilters';
 import { patientVaccineRoutes } from './patient/patientVaccine';
 import { patientDocumentMetadataRoutes } from './patient/patientDocumentMetadata';
 
+import { patientDeath } from './patient/patientDeath';
 import { patientProfilePicture } from './patient/patientProfilePicture';
 import { activeCovid19PatientsHandler } from '../../routeHandlers';
 
@@ -138,6 +141,7 @@ patientRelations.get(
 );
 
 patientRelations.use(patientProfilePicture);
+patientRelations.use(patientDeath);
 
 patientRelations.get(
   '/:id/referrals',
@@ -307,7 +311,7 @@ patientRoute.get(
       offset: page * rowsPerPage,
     });
 
-    const count = lastEncounterMedications.count;
+    const { count } = lastEncounterMedications;
     const data = lastEncounterMedications.rows.map(x => x.forResponse());
 
     res.send({
@@ -418,6 +422,13 @@ patientRoute.get(
           patients.*,
           encounters.id AS encounter_id,
           encounters.encounter_type,
+          CASE
+            WHEN patients.date_of_death IS NOT NULL THEN 'deceased'
+            WHEN encounters.encounter_type = 'emergency' THEN 'emergency'
+            WHEN encounters.encounter_type = 'clinic' THEN 'outpatient'
+            WHEN encounters.encounter_type IS NOT NULL THEN 'inpatient'
+            ELSE NULL
+          END AS patient_status,
           department.id AS department_id,
           department.name AS department_name,
           location.id AS location_id,
@@ -455,3 +466,103 @@ patientRoute.get('/program/activeCovid19Patients', asyncHandler(activeCovid19Pat
 
 patientRoute.use(patientVaccineRoutes);
 patientRoute.use(patientDocumentMetadataRoutes);
+
+/**
+ * Helper function and endpoints for patient additional data.
+ *
+ * Looks up selected field in AdditionalData records and
+ * uses a configured questionIds as a fallback
+ *
+ * The ideal way to get this data would be to allow survey questions to be configured
+ * such that they write their answers to patient record fields. However in the meantime
+ * these endpoint handlers allow easy and consistent access to the data on the front end
+ */
+async function getPatientAdditionalData(req, field, questionId) {
+  const {
+    params,
+    models: { Patient, PatientAdditionalData },
+  } = req;
+  const patientId = params.id;
+
+  const patient = await Patient.findByPk(patientId);
+  if (!patient) throw new NotFoundError();
+
+  const patientAdditionalData = await PatientAdditionalData.findOne({
+    where: { patientId: patient.id },
+    include: PatientAdditionalData.getFullReferenceAssociations(),
+  });
+
+  const value = patientAdditionalData?.dataValues[field];
+
+  if (value) {
+    return value;
+  }
+
+  const result = await req.db.query(
+    `SELECT body, survey_responses.end_time
+       FROM survey_response_answers
+       LEFT JOIN survey_responses
+        ON (survey_responses.id = survey_response_answers.response_id)
+       LEFT JOIN encounters
+        ON (survey_responses.encounter_id = encounters.id)
+       WHERE
+          data_element_id = :questionId
+        AND
+          encounters.patient_id = :patientId`,
+    {
+      replacements: {
+        patientId,
+        questionId,
+      },
+      type: QueryTypes.SELECT,
+    },
+  );
+
+  const resultsWithAnswers = result
+    .filter(({ body }) => !!body)
+    .sort(({ end_time: endTime1 }, { end_time: endTime2 }) => moment(endTime1).isAfter(endTime2));
+
+  if (resultsWithAnswers.length === 0) {
+    return '';
+  }
+
+  return resultsWithAnswers[0].body;
+}
+
+patientRoute.get(
+  '/:id/passportNumber',
+  asyncHandler(async (req, res) => {
+    const questionId = config?.questionCodeIds?.passportNumber;
+
+    if (!questionId) {
+      res.send('');
+      return;
+    }
+
+    req.checkPermission('read', 'Patient');
+
+    const value = await getPatientAdditionalData(req, 'passport', questionId);
+    res.json(value);
+  }),
+);
+
+patientRoute.get(
+  '/:id/nationality',
+  asyncHandler(async (req, res) => {
+    const questionId = config?.questionCodeIds?.citizenship;
+    if (!questionId) {
+      res.send('');
+      return;
+    }
+
+    const value = await getPatientAdditionalData(req, 'nationalityId', questionId);
+
+    if (!value) {
+      res.send('');
+      return;
+    }
+
+    const record = await req.models.ReferenceData.findByPk(value);
+    res.json(record?.dataValues?.name);
+  }),
+);