+++ packages/shared-src/src/reports/reportDefinitions.js
@@ -79,9 +79,19 @@ export const REPORT_DEFINITIONS = [
   },
   {
     name: 'COVID-19 Tests - Line list',
-    id: 'covid-swab-lab-test-list',
+    id: 'fiji-covid-swab-lab-test-list',
     allFacilities: true,
     parameters: [{ parameterField: 'VillageField' }, { parameterField: 'LabTestLaboratoryField' }],
+  },
+  {
+    name: 'Palau COVID-19 Test - Line list',
+    id: 'palau-covid-swab-lab-test-list',
+    allFacilities: true,
+  },
+  {
+    name: 'Kiribati COVID-19 Test - Line list',
+    id: 'kiribati-covid-swab-lab-test-list',
+    allFacilities: true,
   },
   {
     name: 'COVID-19 Tests - Summary',