+++ CHANGELOG.md
@@ -12,6 +12,98 @@
 
 ### Tweaks ⚖️
 
+- (left blank)
+
+### Bug fixes 🐛
+
+- Fixed temp folders not already existing on Windows hosts in lan (was already fixed in sync)
+
+### Infrastructure and maintenance 🛠
+
+- (left blank)
+
+## v1.12.3
+
+### Manual steps required after upgrade ⚙
+
+- Consider adding `kiribati-covid-swab-lab-test-list` to disabledReports.
+
+### Features ⭐
+
+- New report for Kiribati.
+
+### Tweaks ⚖️
+
+- (left blank)
+
+### Bug fixes 🐛
+
+- (left blank)
+
+### Infrastructure and maintenance 🛠
+
+- (left blank)
+
+## v1.12.2
+
+### Manual steps required after upgrade ⚙
+
+- (left blank)
+
+### Features ⭐
+
+- (left blank)
+
+### Tweaks ⚖️
+
+- (left blank)
+
+### Bug fixes 🐛
+
+- (left blank)
+
+### Infrastructure and maintenance 🛠
+
+- (left blank)
+
+## v1.12.1
+
+### Manual steps required after upgrade ⚙
+
+- change any reference to `covid-swab-lab-test-list` (such as in disabledReports) to `fiji-covid-swab-lab-test-list`
+
+### Features ⭐
+
+- (left blank)
+
+### Tweaks ⚖️
+
+- (left blank)
+
+### Bug fixes 🐛
+
+- (left blank)
+
+### Infrastructure and maintenance 🛠
+
+- (left blank)
+
+## v1.12.0
+
+### Manual steps required after upgrade ⚙
+
+- Run `yarn` to install new dependency (check-disk-usage) and downgrade babel-jest
+- Add the asset `certificate-bottom-half-img` to each deployment
+- Reimport facilities after mobile update has been deployed
+
+### Features ⭐
+
+- Central server has the ability to check available disk space based on current config.
+- Lan servers will auto-discontinue medications after their end date is due (based on facility).
+- Replace the bottom half of certificates with a configurable image
+
+### Tweaks ⚖️
+
 - Add `react-hooks/exhaustive-deps` and `react-hooks/rules-of-hooks` linter rules
 
 ### Bug fixes 🐛
@@ -92,7 +184,7 @@
 
 ### Tweaks ⚖️
 
-- (left blank)
+- Updated logic for the covid-swab-lab-test-list report
 
 ### Bug fixes 🐛